Change tuple and matrix to use float instead of double

+++ libraries/Utility/include/Matrix.hpp
@@ -17,38 +17,38 @@ namespace utility {
 template <uint8_t rows, uint8_t cols> 
 struct alignas(32) Matrix
 {
-  std::array<double, rows*cols> data;
+  std::array<float, rows*cols> data;
   template<typename... T>
-  explicit Matrix(T... args) noexcept: data{static_cast<double>(args)...}{};
-  explicit Matrix(const std::array<double, rows*cols>& arr) noexcept : data(arr) {}
+  explicit Matrix(T... args) noexcept: data{static_cast<float>(args)...}{};
+  explicit Matrix(const std::array<float, rows*cols>& arr) noexcept : data(arr) {}
 
   bool operator==(const Matrix& rhs) const noexcept;
 
-  const double& at(const std::size_t& row, const std::size_t& column) const noexcept;
-  double& at(const std::size_t& row, const std::size_t& column) noexcept;
+  const float& at(const std::size_t& row, const std::size_t& column) const noexcept;
+  float& at(const std::size_t& row, const std::size_t& column) noexcept;
 
   static Matrix<rows, cols> identity() noexcept;
 
   Matrix<rows, cols>& operator*=(const Matrix<rows, cols>& rhs) noexcept;
   Matrix<rows, cols> transpose() const noexcept;
-  double determinant() const noexcept;
+  float determinant() const noexcept;
   bool invertible() const noexcept;
 };
 
 template <uint8_t rows, uint8_t cols>
-inline const double& Matrix<rows, cols>::at(const std::size_t& row, const std::size_t& column) const noexcept{
+inline const float& Matrix<rows, cols>::at(const std::size_t& row, const std::size_t& column) const noexcept{
   return data[row*cols+column];
 }
 
 template <uint8_t rows, uint8_t cols> 
-inline double& Matrix<rows, cols>::at(const std::size_t& row, const std::size_t& column) noexcept{
-  return const_cast<double&>(const_cast<const Matrix*>(this)->at(row, column));
+inline float& Matrix<rows, cols>::at(const std::size_t& row, const std::size_t& column) noexcept{
+  return const_cast<float&>(const_cast<const Matrix*>(this)->at(row, column));
 }
 
 template <uint8_t rows, uint8_t cols> 
 bool Matrix<rows, cols>::operator==(const Matrix& rhs) const noexcept{
   for (size_t i = 0; i < data.size(); ++i) {
-    if (!floatNearlyEqual<double>(data[i], rhs.data[i])) {
+    if (!floatNearlyEqual<float>(data[i], rhs.data[i])) {
       return false;
     }
   }
@@ -82,10 +82,10 @@ constexpr typename std::enable_if<rows == 4, Tuple>::type
 operator*(const Matrix<rows, cols>& lhs, const Tuple& rhs) noexcept {
   Tuple result;
   
-  const double* matrix_data = lhs.data.data();
+  const float* matrix_data = lhs.data.data();
   
   for (int i = 0; i < 4; ++i) {
-    double sum = 0.0;
+    float sum = 0.0;
     for (int j = 0; j < 4; ++j) {
       sum += matrix_data[i * 4 + j] * (&rhs.x)[j];
     }
@@ -130,21 +130,21 @@ Matrix<rows, cols> Matrix<rows, cols>::transpose() const noexcept{
 }
 
 template<>
-inline double Matrix<1,1>::determinant() const noexcept{
+inline float Matrix<1,1>::determinant() const noexcept{
   auto determinant = this->at(0,0);
 
   return determinant;
 }
 
 template<>
-inline double Matrix<2,2>::determinant() const noexcept{
+inline float Matrix<2,2>::determinant() const noexcept{
   auto determinant = this->at(0,0)*this->at(1,1) - this->at(0,1)*this->at(1,0);
 
   return determinant;
 }
 
 template<>
-inline double Matrix<3,3>::determinant() const noexcept{
+inline float Matrix<3,3>::determinant() const noexcept{
   return
     + this->at(0,0) * (this->at(1,1) * this->at(2,2) - this->at(2,1) * this->at(1,2))
     - this->at(1,0) * (this->at(0,1) * this->at(2,2) - this->at(2,1) * this->at(0,2))
@@ -152,13 +152,13 @@ inline double Matrix<3,3>::determinant() const noexcept{
 }
 
 template<>
-inline double Matrix<4,4>::determinant() const noexcept{
-  const double SubFactor00 = this->at(2,2) * this->at(3,3) - this->at(3,2) * this->at(2,3);
-  const double SubFactor01 = this->at(2,1) * this->at(3,3) - this->at(3,1) * this->at(2,3);
-  const double SubFactor02 = this->at(2,1) * this->at(3,2) - this->at(3,1) * this->at(2,2);
-  const double SubFactor03 = this->at(2,0) * this->at(3,3) - this->at(3,0) * this->at(2,3);
-  const double SubFactor04 = this->at(2,0) * this->at(3,2) - this->at(3,0) * this->at(2,2);
-  const double SubFactor05 = this->at(2,0) * this->at(3,1) - this->at(3,0) * this->at(2,1);
+inline float Matrix<4,4>::determinant() const noexcept{
+  const float SubFactor00 = this->at(2,2) * this->at(3,3) - this->at(3,2) * this->at(2,3);
+  const float SubFactor01 = this->at(2,1) * this->at(3,3) - this->at(3,1) * this->at(2,3);
+  const float SubFactor02 = this->at(2,1) * this->at(3,2) - this->at(3,1) * this->at(2,2);
+  const float SubFactor03 = this->at(2,0) * this->at(3,3) - this->at(3,0) * this->at(2,3);
+  const float SubFactor04 = this->at(2,0) * this->at(3,2) - this->at(3,0) * this->at(2,2);
+  const float SubFactor05 = this->at(2,0) * this->at(3,1) - this->at(3,0) * this->at(2,1);
 
   const Tuple determinantCoefficients{
     + (this->at(1,1) * SubFactor00 - this->at(1,2) * SubFactor01 + this->at(1,3) * SubFactor02),
@@ -173,8 +173,8 @@ inline double Matrix<4,4>::determinant() const noexcept{
 }
 
 template <uint8_t rows, uint8_t cols>
-double Matrix<rows,cols>::determinant() const noexcept{
-  double determinant{};
+float Matrix<rows,cols>::determinant() const noexcept{
+  float determinant{};
 
   for(std::size_t colIndex{0}; colIndex < cols; colIndex++){
     determinant += this->at(0, colIndex)*cofactor(*this, 0, colIndex);
@@ -204,14 +204,14 @@ inline submatrix(const Matrix<rows, cols>& matrix, const std::size_t& skipRow, c
 }
 
 template<uint8_t rows, uint8_t cols>
-double minor(const Matrix<rows, cols>& matrix, const std::size_t& row, const std::size_t& column) noexcept{
+float minor(const Matrix<rows, cols>& matrix, const std::size_t& row, const std::size_t& column) noexcept{
     auto subMatrix =  submatrix(matrix, row, column);
     return subMatrix.determinant();  
 }
 
 template<uint8_t rows, uint8_t cols>
-inline double cofactor(const Matrix<rows, cols>& matrix, const std::size_t& row, const std::size_t& column) noexcept{
-  double cofactor = (row+column)%2 ? -1*minor(matrix, row, column) : minor(matrix, row, column);
+inline float cofactor(const Matrix<rows, cols>& matrix, const std::size_t& row, const std::size_t& column) noexcept{
+  float cofactor = (row+column)%2 ? -1*minor(matrix, row, column) : minor(matrix, row, column);
 
   return cofactor;
 }
@@ -225,29 +225,29 @@ bool Matrix<rows, cols>::invertible() const noexcept{
 template<uint8_t rows, uint8_t cols>
 typename std::enable_if<(rows==4 && cols==4), Matrix<rows, cols>>::type
 inline inverse(const Matrix<rows,cols>& matrix) noexcept{
-  const double Coef00 = matrix.at(2,2) * matrix.at(3,3) - matrix.at(3,2) * matrix.at(2,3);
-  const double Coef02 = matrix.at(1,2) * matrix.at(3,3) - matrix.at(3,2) * matrix.at(1,3);
-  const double Coef03 = matrix.at(1,2) * matrix.at(2,3) - matrix.at(2,2) * matrix.at(1,3);
+  const float Coef00 = matrix.at(2,2) * matrix.at(3,3) - matrix.at(3,2) * matrix.at(2,3);
+  const float Coef02 = matrix.at(1,2) * matrix.at(3,3) - matrix.at(3,2) * matrix.at(1,3);
+  const float Coef03 = matrix.at(1,2) * matrix.at(2,3) - matrix.at(2,2) * matrix.at(1,3);
 
-  const double Coef04 = matrix.at(2,1) * matrix.at(3,3) - matrix.at(3,1) * matrix.at(2,3);
-  const double Coef06 = matrix.at(1,1) * matrix.at(3,3) - matrix.at(3,1) * matrix.at(1,3);
-  const double Coef07 = matrix.at(1,1) * matrix.at(2,3) - matrix.at(2,1) * matrix.at(1,3);
+  const float Coef04 = matrix.at(2,1) * matrix.at(3,3) - matrix.at(3,1) * matrix.at(2,3);
+  const float Coef06 = matrix.at(1,1) * matrix.at(3,3) - matrix.at(3,1) * matrix.at(1,3);
+  const float Coef07 = matrix.at(1,1) * matrix.at(2,3) - matrix.at(2,1) * matrix.at(1,3);
 
-  const double Coef08 = matrix.at(2,1) * matrix.at(3,2) - matrix.at(3,1) * matrix.at(2,2);
-  const double Coef10 = matrix.at(1,1) * matrix.at(3,2) - matrix.at(3,1) * matrix.at(1,2);
-  const double Coef11 = matrix.at(1,1) * matrix.at(2,2) - matrix.at(2,1) * matrix.at(1,2);
+  const float Coef08 = matrix.at(2,1) * matrix.at(3,2) - matrix.at(3,1) * matrix.at(2,2);
+  const float Coef10 = matrix.at(1,1) * matrix.at(3,2) - matrix.at(3,1) * matrix.at(1,2);
+  const float Coef11 = matrix.at(1,1) * matrix.at(2,2) - matrix.at(2,1) * matrix.at(1,2);
 
-  const double Coef12 = matrix.at(2,0) * matrix.at(3,3) - matrix.at(3,0) * matrix.at(2,3);
-  const double Coef14 = matrix.at(1,0) * matrix.at(3,3) - matrix.at(3,0) * matrix.at(1,3);
-  const double Coef15 = matrix.at(1,0) * matrix.at(2,3) - matrix.at(2,0) * matrix.at(1,3);
+  const float Coef12 = matrix.at(2,0) * matrix.at(3,3) - matrix.at(3,0) * matrix.at(2,3);
+  const float Coef14 = matrix.at(1,0) * matrix.at(3,3) - matrix.at(3,0) * matrix.at(1,3);
+  const float Coef15 = matrix.at(1,0) * matrix.at(2,3) - matrix.at(2,0) * matrix.at(1,3);
 
-  const double Coef16 = matrix.at(2,0) * matrix.at(3,2) - matrix.at(3,0) * matrix.at(2,2);
-  const double Coef18 = matrix.at(1,0) * matrix.at(3,2) - matrix.at(3,0) * matrix.at(1,2);
-  const double Coef19 = matrix.at(1,0) * matrix.at(2,2) - matrix.at(2,0) * matrix.at(1,2);
+  const float Coef16 = matrix.at(2,0) * matrix.at(3,2) - matrix.at(3,0) * matrix.at(2,2);
+  const float Coef18 = matrix.at(1,0) * matrix.at(3,2) - matrix.at(3,0) * matrix.at(1,2);
+  const float Coef19 = matrix.at(1,0) * matrix.at(2,2) - matrix.at(2,0) * matrix.at(1,2);
 
-  const double Coef20 = matrix.at(2,0) * matrix.at(3,1) - matrix.at(3,0) * matrix.at(2,1);
-  const double Coef22 = matrix.at(1,0) * matrix.at(3,1) - matrix.at(3,0) * matrix.at(1,1);
-  const double Coef23 = matrix.at(1,0) * matrix.at(2,1) - matrix.at(2,0) * matrix.at(1,1);
+  const float Coef20 = matrix.at(2,0) * matrix.at(3,1) - matrix.at(3,0) * matrix.at(2,1);
+  const float Coef22 = matrix.at(1,0) * matrix.at(3,1) - matrix.at(3,0) * matrix.at(1,1);
+  const float Coef23 = matrix.at(1,0) * matrix.at(2,1) - matrix.at(2,0) * matrix.at(1,1);
 
   const Tuple Fac0{Coef00, Coef00, Coef02, Coef03};
   const Tuple Fac1{Coef04, Coef04, Coef06, Coef07};
@@ -264,7 +264,7 @@ inline inverse(const Matrix<rows,cols>& matrix) noexcept{
   const Tuple SignA{+1, -1, +1, -1};
   const Tuple SignB{-1, +1, -1, +1};
 
-  const double oneOverDeterminant = static_cast<double>(1)/matrix.determinant();
+  const float oneOverDeterminant = static_cast<float>(1)/matrix.determinant();
 
   const Tuple Inv0{(Vec1 * Fac0 - Vec2 * Fac1 + Vec3 * Fac2) * oneOverDeterminant * SignA};
   const Tuple Inv1{(Vec0 * Fac0 - Vec2 * Fac3 + Vec3 * Fac4) * oneOverDeterminant * SignB};
